Moving specific flask app related config options to the server side
